# docs/development/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MomentRead local API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4317"
    }
  ],
  "paths": {
...
    "/api/runtime": {
      "get": {
        "operationId": "get__runtime",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "413": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "502": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/books": {
      "get": {
        "operationId": "get__books",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookListResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "413": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "502": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__books",
        "parameters": [],
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/books"
                }
              }
            }
          },
          "400": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "413": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "502": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        }
      }
    },
    "/api/books/{id}": {
      "patch": {
        "operationId": "patch__books_id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/books"
                }
              }
            }
          },
          "400": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "413": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "502": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookMetadataPatchSchema"
              }
            }
          }
        }
      }
    },
    "/api/books/{id}/state": {
      "get": {
        "operationId": "get__books_id_state",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookStateResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "413": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "502": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/books/{id}/workspace": {
      "patch": {
        "operationId": "patch__books_id_workspace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/workspaces"
                }
              }
            }
          },
          "400": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "409": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "413": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "429": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "500": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "502": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          },
          "503": {
            "description": "失败；错误码、可操作说明、可重试性与请求 ID。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspacePatchSchema"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookMetadataPatchSchema": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "author": {
            "type": "string",
            "maxLength": 500
          },
          "language": {
            "type": "string",
            "maxLength": 100
          },
          "translator": {
            "type": "string",
            "maxLength": 500
          },
          "edition": {
            "type": "string",
            "maxLength": 1000
          },
          "identifier": {
            "type": "string",
            "maxLength": 500
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponseSchema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "additionalProperties": false
          },
          "requestId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "error",
          "requestId"
        ],
        "additionalProperties": false
      },
...
      "RuntimeResponseSchema": {
        "type": "object",
        "properties": {
          "installed": {
            "type": "boolean"
          },
          "version": {
            "type": [
              "string",
              "null"
            ]
          },
          "authReported": {
            "type": "boolean"
          },
          "invocationVerified": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "installed",
          "version",
          "authReported",
          "invocationVerified",
          "message"
        ],
        "additionalProperties": false
      },
...
    }
  }
}